Falling objects + random colors
Fixed problems relating to the creation and falling of the rocks and gems, as well as problems afflicting the ability to randomize colors of gems and rocks. Also added start of player code, as of right now does not work, may have to move bits of code to program to allow it to function.

Still need to do
Finish player code
Add collision
Reduce spawning of gems + rocks (way to fast)